refactor(web): isolate agent config tab
